requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
pillow==11.0.0
requests==2.32.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
